Fix parsing for lines ending with a space.
Lines can have a space, but they are discarded. This would mean the position would be wrong after the rtrim, and a line would be skipped by the reg-exp. I'm not sure the characters that are stripped are actually there, but to be as close to the current code I have just removed the space character.
Testdata (MT940) can be send upon request.