Fix for bookdates before valuedates around januari 1st.
The compasion of the year of the dates was moot as they were set to equal a few lines earlier. This resulted in wrong bookdates.

By moving the setTime line directly after creating the book date, the comparison of both \DateTime objects can be used and will properly update the bookdate if it actually is before the value date.